COMP: Rename helper script and make sure the script is installed.
* Later, the organization of the file will be more 'pythonic'. See http://na-mic.org/Mantis/view.php?id=1365
